+++ example/lib/config/environment.dart
@@ -1,0 +1,4 @@
+class EnvironmentConfig {
+  static String bridgefyColor = '0x00FF4040';
+  static String apiKey = "YOUR_API_KEY_HERE";
+}