+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Bridgefy Flutter 1.1.4
+# Bridgefy Flutter 1.1.5
 ### BLUETOOTH MESH NETWORKS
 
 Devices running your app create Bluetooth Low-Energy mesh networks, which work in large crowds, during and after natural disasters, and whenever else your users lose access to the Internet.
@@ -13,4 +13,13 @@ is fragile and unstable.
 All of your users participate in the mesh networks without having to provide any extra permissions, perform any actions, or even be aware of participating in traffic. Battery, storage, and processor are 
 all taken
 
+## Change Log
 
+### Version 1.1.5
+
+* Fix: Use maxSdkVersion 34 for android.permission.ACCESS_FINE_LOCATION
+* Chore: Add example config
+* Chore: Add log and message entities
+* Chore: Add chat and SDK providers
+* Chore: Add provider and upgrade permission_handler plugins
+* Docs: Update README example with the new path to use your API key